feat(api): change long to int

+++ championship-manager-api/championship-manager-application/src/main/kotlin/fr/pk/championshipmanagerapplication/dto/ChampionnatDto.kt
@@ -1,3 +1,3 @@
 package fr.pk.championshipmanagerapplication.dto
 
-data class ChampionnatDto(val id: Long?, val nom: String)
+data class ChampionnatDto(val id: Int?, val nom: String)
